pyproject.toml
[tool.poetry]
name = "usc_lr_timer"
version = "0.1.0"
description = "USC Law Review Timer"
authors = ["Perry Vargas <perrybvargas@gmail.com>"]

[tool.poetry.dependencies]
python = "3.7"
pyside2 = "^5.15.0"
google-api-python-client = "^1.10.0"
google-auth-httplib2 = "^0.0.4"
google-auth-oauthlib = "^0.4.1"
isort = "^5.3.2"

[tool.poetry.dev-dependencies]
pytest = "^6.0"
ipython = "^7.17.0"
briefcase = "^0.3.3"
pytest-bdd = "^3.4.0"
pytest-qt = "^3.3.0"
pytest-mock = "^3.2.0"
black = "^19.10b0"
pytest-cov = "^2.10.0"
freezegun = "^0.3.15"
coveralls = "^2.1.1"
invoke = "^1.4.1"
flake8 = "^3.8.3"

[tool.briefcase]
project_name = "USC Law Review Timer"
bundle = "ult.example"
version = "1.0.0"
url = "https://example.com/ult"
license = "LGPL"
author = 'Perry Vargas'
author_email = "perrybvargas@gmail.com"

[tool.briefcase.app.usc_lr_timer]
formal_name = "USC Law Review Timer"
description = "Testing if briefcase can freeze a scientific qt app"
icon = "usc_lr_timer/resources/icon"
sources = ['usc_lr_timer']
requires = [
    "pyside2==5.15.0",
    "google-api-python-client==1.10.0",
    "google-auth-httplib2==0.0.4",
    "google-auth-oauthlib==0.4.1",
]


[tool.briefcase.app.usc_lr_timer.macOS]
requires = []

[tool.briefcase.app.usc_lr_timer.linux]
requires = []

[tool.briefcase.app.usc_lr_timer.windows]
requires = []

# Mobile deployments
[tool.briefcase.app.usc_lr_timer.iOS]
requires = []

[tool.briefcase.app.usc_lr_timer.android]
requires = []

[tool.pytest.ini_options]
bdd_features_base_dir = "tests/bdd/features/"

[tool.coverage.report]
exclude_lines = [
    "if __name__ == .__main__.:",
]
omit = ["usc_lr_timer/__main__.py"]

[tool.black]
line-length = 80
target-version = ['py37']
skip-string-normalization = true

[tool.isort]
profile = "black"
line_length=80
force_alphabetical_sort_within_sections = true
src_paths = ['usc_lr_timer', 'tests']
force_sort_within_sections = true

[build-system]
requires = [
    "poetry>=0.12",
    "briefcase"
]
build-backend = "poetry.masonry.api"
